video-creator-app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
moviepy==1.0.3
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
pillow==10.1.0
aiofiles==23.2.1
